seperated out the test cases from lib.rs and then made adjustment so that all the tests would work as expected
